Refactor processors into folder.
